+++ CHANGELOG.md
@@ -6,3 +6,13 @@
     - shows entropy
     
     - can pass number of last commits to analyze using --last option
+
+## 0.0.3
+- Added Phase 2: Metrics Calculation
+    - by default analyzes last 50 commits
+    - shows top 10 most changed files default
+    - shows total files changed
+    - shows entropy calcutaion
+    
+    - use dart run bin/main.dart -h to know about the cmds if running locally
+    - use rea -h to know about the cmds if installed globally